wec2xls/requirements.txt
ndjson==0.3.1
python-dateutil==2.8.1
six==1.15.0
XlsxWriter==1.3.7
